Fix guest agent fields messaging
